some changes

+++ backend/Controllers/OTPController.js
@@ -1,8 +1,10 @@
-const OTP= require('../models/otpModel')
-const {generateOTP} = require('../util/generateOTP')
+const OTP = require('../models/otpModel')
+const User = require('../models/UserModel');
+const { generateOTP } = require('../util/generateOTP')
 const sendEmail = require('../util/sendEmail')
 
-const {AUTH_EMAIL}=process.env
+
+const { AUTH_EMAIL } = process.env
 
 
 const verifyOTP = async ({ email, otp }) => {
@@ -23,46 +25,48 @@ const verifyOTP = async ({ email, otp }) => {
         if (otp !== otpStored) {
             return "Invalid OTP"
         }
+        const existingUser = User.findOne({email})
+        existingUser.verified = true;
         return true;
     } catch (err) {
         throw err;
     }
 };
 
-const deleteOTP = (email)=>{
+const deleteOTP = (email) => {
     try {
-        OTP.deleteOne({email})
-    }catch (err){throw err}
+        OTP.deleteOne({ email })
+    } catch (err) { throw err }
 }
 
-const sendOTP = async ({email,subject,message,duration = 1})=>{
+const sendOTP = async ({ email }) => {
     try {
-        if (!{email,subject,message,duration}){
-            throw Error("Provide values for email ,subject,message")
-        }
+        if (!email) {return "Provide value for email "}
+        const existingUser = await User.findOne({email})
+        if (!existingUser){return "There is no account for this email"}
         //clear old records
-        await OTP.deleteOne({email})
+        await OTP.deleteOne({ email })
         //generate Pin
         const generatedOTP = generateOTP()
         //send email
         const mailOptions = {
-            from : AUTH_EMAIL,
-            to : email,
-            subject,
-            html :
-            `
-            <p>${message}</p>
+            from: AUTH_EMAIL,
+            to: email,
+            subject: "Email verification",
+            html:
+                `
+            <p>verify your email with the code below : </p>
                 <p>Your OTP is: <strong>${generatedOTP}</strong></p>
-                <p>This OTP will expire in ${duration} ${duration===1 ? 'hour': 'hours'}.</p>
+                <p>This OTP will expire in 1 hour.</p>
             `
         }
         await sendEmail(mailOptions)
         //save OTP IN DATABSE
-        const newOTP =  new OTP({
+        const newOTP = new OTP({
             email,
             otp: generatedOTP,
-            createdAt : Date.now(),
-            expiredAt : Date.now() + 3600000 * +duration,
+            createdAt: Date.now(),
+            expiredAt: Date.now() + 3600000,
         })
         const createdOTPRecord = await newOTP.save()
         return createdOTPRecord
@@ -70,4 +74,4 @@ const sendOTP = async ({email,subject,message,duration = 1})=>{
         throw error
     }
 }
-module.exports = {sendOTP,verifyOTP,deleteOTP}
+module.exports = { sendOTP, verifyOTP, deleteOTP }
